backend/requirement.txt
fastapi
uvicorn
python-dotenv
opencv-python
scikit-learn
face-recognition
imutils
